Nearly Compile Ready
Reduced errors from 7284 to around 200.
Biggest change is rebuilding all the obj references for shaders, updating old d3d and dx refs to bgfx.
Still not compiling through cmake, just a few things left to ref to. Once it comiles i can try out the game and at least see if it will load.

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WW3D2/render2dsentence.cpp
@@ -40,6 +40,7 @@
 #include "wwprofile.h"
 #include "wwmemlog.h"
 #include "BGFXWrapper.h" // Ported from dx8wrapper.h to BGFXWrapper.h
+#include "ww3d.h"
 
 #ifdef _INTERNAL
 // for occasional debugging...
@@ -1720,8 +1721,8 @@ FontCharsClass::Grow_Unicode_Array (WCHAR ch)
 		return ;
 	} 
 
-	uint16 first_index	= min( FirstUnicodeChar, ch );
-	uint16 last_index		= max( LastUnicodeChar, ch );
+	uint16 first_index	= min( (uint16)FirstUnicodeChar, (uint16)ch );
+	uint16 last_index		= max( (uint16)LastUnicodeChar, (uint16)ch );
 	uint16 count			= (last_index - first_index) + 1;
 
 	//
